include by viewModels() pattern and rxJava for Search, Popular and topMovie feature

+++ app/src/main/java/com/sicoapp/movieapp/ui/movie/crew/CrewMovieFragment.kt
@@ -33,8 +33,8 @@ class CrewMovieFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
 
         super.onCreate(savedInstanceState)
-        arguments?.getInt(CREW_ID, -1)?.let {
-            crewId = it
+        arguments?.getLong(CREW_ID, -1)?.let {
+            crewId = it.toInt()
         }
     }
